requirements.txt
psycopg2==2.8.5
django==3.0.5
Pillow==7.1.2
django-parler==2.0.1